+++ docs/changelog.md
@@ -246,7 +246,7 @@
 
 #### num::prob::rand
 - new trait `RandTry`.
-- new types `SplitMix64`, `StdRand`.
+- new types `RandQualities`, `SplitMix64`, `StdRand`.
 - remove the traits: `RandAlloc`, `RandStd`.
 - rename `define_pcg!` to `rand_pcg!`
 - rename `define_xorshift!` to `rand_xorshift!`.